feat(childVisibilityAndCDN): Register nav visibility methods, and use CDN
Closes #522
Closes #520

This lets us register methods to be used for `visibility` and
`childVisibility` values in nav menus. Because we're now loading some
nav menus from JSON, we can't put arbitrary functions in anymore.

Now we can register new methods with `rxVisibility`, and reference them
in the nav as:

```
visibility: { methodName: 'someRegisteredName', args: { foo: 'bar' } }
```

We also start using the correct CDN file, checking whether or not
we're in a production setting